#142 - rename projects table file and update test files
